IN-47 feat: change paths for schemas & types
