+++ src/features/auth/lib/types/auth.types.ts
@@ -1,0 +1,1 @@
+// типы для auth
